requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
pyyaml==6.0.1
google-genai==0.1.0
twilio==8.10.0
